requirements.txt
requests
PyYAML
influxdb-client
python-dotenv
Flask
